2020/4/30

+++ src/main/java/q/community/community/controller/PublishController.java
@@ -8,12 +8,10 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import q.community.community.dto.QuestionDTO;
-import q.community.community.mapper.QuestionMapper;
 import q.community.community.model.Question;
 import q.community.community.model.User;
 import q.community.community.service.QuestionService;
 
-import javax.jws.WebParam;
 import javax.servlet.http.HttpServletRequest;
 
 @Controller
@@ -23,7 +21,7 @@ public class PublishController {
     private QuestionService questionService;
 
     @GetMapping("/publish/{id}")
-    public String edit(@PathVariable(name="id")Integer id,
+    public String edit(@PathVariable(name="id")Long id,
                        Model model){
         QuestionDTO question = questionService.getById(id);
 
@@ -46,7 +44,7 @@ public class PublishController {
             @RequestParam(value = "title",required = false) String title,
             @RequestParam(value = "description" ,required = false) String description,
             @RequestParam(value = "tag",required = false) String tag,
-            @RequestParam(value = "id",required = false) Integer id,
+            @RequestParam(value = "id",required = false) Long id,
             HttpServletRequest request,
             Model model
             ){
